re-structure files

+++ app/src/main/java/com/unknown/sdust/jwgl_tp/data/VersionData.java
@@ -1,9 +1,9 @@
 package com.unknown.sdust.jwgl_tp.data;
 
 public class VersionData {
-    public long login_lastVersion;
-    public long table_lastVersion;
-    public VersionData(long l, long t){
+    public int login_lastVersion;
+    public int table_lastVersion;
+    public VersionData(int l, int t){
         login_lastVersion = l;
         table_lastVersion = t;
     }
